+++ src/main/java/frc/robot/subsystems/drive/MAXSwerveModule.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.util.rev;
+package frc.robot.subsystems.drive;
 
 import org.littletonrobotics.junction.AutoLogOutput;
 
@@ -13,8 +13,7 @@ import edu.wpi.first.math.kinematics.SwerveModulePosition;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
 import frc.robot.util.Constants.FieldConstants;
 import frc.robot.util.Constants.ModuleConstants;
-import monologue.Logged;
-import monologue.Annotations.Log;
+import frc.robot.util.rev.Neo;
 
 public class MAXSwerveModule implements MAXSwerveModuleIO {
     private final Neo driveMotor;
