+++ src/main/java/frc/robot/subsystems/drive/MAXSwerveModuleIO.java
@@ -1,4 +1,4 @@
-package frc.robot.util.rev;
+package frc.robot.subsystems.drive;
 
 import org.littletonrobotics.junction.AutoLog;
