+++ src/main/java/frc/robot/util/auto/PathPlannerStorage.java
@@ -22,8 +22,8 @@ import frc.robot.util.custom.PatriSendableChooser;
 import frc.robot.Robot;
 import frc.robot.RobotContainer;
 import frc.robot.Robot.GameMode;
-import frc.robot.subsystems.Limelight;
-import frc.robot.subsystems.Swerve;
+import frc.robot.subsystems.drive.Swerve;
+import frc.robot.subsystems.vision.Limelight;
 import monologue.Logged;
 import monologue.Monologue;
 import monologue.Annotations.IgnoreLogged;
